fix: Remove Timer's CacheTracker as it may fail to start Timer
This was originally added to allow for the cache to fill up before running tasks from the Timer extension in case they needed them. However, the way the cache fills up is too finicky to reliably track and so this is being removed. Any tasks that did use the cache on a task that might run on start should opt instead to use the hikari's `rest` API.